Add ability to translate very basic OpenMP
This commit adds functionality in translate.py allowing it to select
compound statements annotated with the "omp parallel" pragma as the
code to be tranlated into OpenCL kernels. An option is added to the
program to explicitly ask for this functionality. Host code is not
properly generated for this case yet.

This commit significantly enhances the code generation functionality of
translate.py and will allow less reliance on pycparser's CGenerator.