Split use cases for TranslationVisitor into methods, allow reuse
The visit method should no longer be directly called outside of the
object itself, now use methods which set up the correct flags for the
specific translation you want. State is no longer initialised in the
constructor and this is done in the methods instead, resetting on each
call so the visitor can be reused.