Make OpenMP mode the default and remove old default behaviour
Previously the program was hardcoded to translate the first function in
its input file into a kernel, and also assumed that this function
contained only a for loop, which was to be parallelised. This behaviour
was still the default. The --omp option has been removed as this is now
the default.
The visit_FileAST method of TranslationVisitor has also been removed as
the visitor should no longer be called on a full AST.